+++ src/vs/platform/agentHost/node/agentService.ts
@@ -3,18 +3,21 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { decodeBase64, VSBuffer } from '../../../base/common/buffer.js';
 import { Emitter } from '../../../base/common/event.js';
 import { Disposable, DisposableStore } from '../../../base/common/lifecycle.js';
 import { observableValue } from '../../../base/common/observable.js';
 import { URI } from '../../../base/common/uri.js';
-import { IFileService } from '../../files/common/files.js';
+import { generateUuid } from '../../../base/common/uuid.js';
+import { FileSystemProviderErrorCode, IFileService, toFileSystemProviderErrorCode } from '../../files/common/files.js';
 import { ILogService } from '../../log/common/log.js';
-import { AgentProvider, AgentSession, IAgent, IAgentCreateSessionConfig, IAgentDescriptor, IAgentService, IAgentSessionMetadata, IAuthenticateParams, IAuthenticateResult, IResourceMetadata } from '../common/agentService.js';
+import { AgentProvider, AgentSession, IAgent, IAgentCreateSessionConfig, IAgentDescriptor, IAgentMessageEvent, IAgentService, IAgentSessionMetadata, IAgentToolCompleteEvent, IAgentToolStartEvent, IAuthenticateParams, IAuthenticateResult, IResourceMetadata } from '../common/agentService.js';
 import { ISessionDataService } from '../common/sessionDataService.js';
 import { ActionType, IActionEnvelope, INotification, ISessionAction } from '../common/state/sessionActions.js';
-import type { IBrowseDirectoryResult, IFetchContentResult, IStateSnapshot, IWriteFileParams, IWriteFileResult } from '../common/state/sessionProtocol.js';
-import { SessionStatus, type ISessionSummary } from '../common/state/sessionState.js';
+import { AhpErrorCodes, AHP_SESSION_NOT_FOUND, ContentEncoding, JSON_RPC_INTERNAL_ERROR, ProtocolError, type IBrowseDirectoryResult, type IDirectoryEntry, type IFetchContentResult, type IStateSnapshot, type IWriteFileParams, type IWriteFileResult } from '../common/state/sessionProtocol.js';
+import { ResponsePartKind, SessionStatus, ToolCallConfirmationReason, ToolCallStatus, TurnState, type IResponsePart, type ISessionSummary, type IToolCallCompletedState, type ITurn } from '../common/state/sessionState.js';
 import { AgentSideEffects } from './agentSideEffects.js';
+import { ISessionDbUriFields, parseSessionDbUri } from './copilot/fileEditTracker.js';
 import { SessionStateManager } from './sessionStateManager.js';
 
 /**
@@ -66,7 +69,7 @@ export class AgentService extends Disposable implements IAgentService {
 			getAgent: session => this._findProviderForSession(session),
 			sessionDataService: this._sessionDataService,
 			agents: this._agents,
-		}, this._logService, this._fileService));
+		}, this._logService));
 	}
 
 	// ---- provider registration ----------------------------------------------
@@ -206,19 +209,125 @@ export class AgentService extends Disposable implements IAgentService {
 	}
 
 	async browseDirectory(uri: URI): Promise<IBrowseDirectoryResult> {
-		return this._sideEffects.handleBrowseDirectory(uri.toString());
+		let stat;
+		try {
+			stat = await this._fileService.resolve(uri);
+		} catch {
+			throw new ProtocolError(AhpErrorCodes.NotFound, `Directory not found: ${uri.toString()}`);
+		}
+
+		if (!stat.isDirectory) {
+			throw new ProtocolError(AhpErrorCodes.NotFound, `Not a directory: ${uri.toString()}`);
+		}
+
+		const entries: IDirectoryEntry[] = (stat.children ?? []).map(child => ({
+			name: child.name,
+			type: child.isDirectory ? 'directory' : 'file',
+		}));
+		return { entries };
 	}
 
 	async restoreSession(session: URI): Promise<void> {
-		return this._sideEffects.handleRestoreSession(session.toString());
+		const sessionStr = session.toString();
+
+		// Already in state manager - nothing to do.
+		if (this._stateManager.getSessionState(sessionStr)) {
+			return;
+		}
+
+		const agent = this._findProviderForSession(session);
+		if (!agent) {
+			throw new ProtocolError(AHP_SESSION_NOT_FOUND, `No agent for session: ${sessionStr}`);
+		}
+
+		// Verify the session actually exists on the backend to avoid
+		// creating phantom sessions for made-up URIs.
+		let allSessions;
+		try {
+			allSessions = await agent.listSessions();
+		} catch (err) {
+			if (err instanceof ProtocolError) {
+				throw err;
+			}
+			const message = err instanceof Error ? err.message : String(err);
+			throw new ProtocolError(JSON_RPC_INTERNAL_ERROR, `Failed to list sessions for ${sessionStr}: ${message}`);
+		}
+		const meta = allSessions.find(s => s.session.toString() === sessionStr);
+		if (!meta) {
+			throw new ProtocolError(AHP_SESSION_NOT_FOUND, `Session not found on backend: ${sessionStr}`);
+		}
+
+		let messages;
+		try {
+			messages = await agent.getSessionMessages(session);
+		} catch (err) {
+			if (err instanceof ProtocolError) {
+				throw err;
+			}
+			const message = err instanceof Error ? err.message : String(err);
+			throw new ProtocolError(JSON_RPC_INTERNAL_ERROR, `Failed to restore session ${sessionStr}: ${message}`);
+		}
+		const turns = this._buildTurnsFromMessages(messages);
+
+		const summary: ISessionSummary = {
+			resource: sessionStr,
+			provider: agent.id,
+			title: meta.summary ?? 'Session',
+			status: SessionStatus.Idle,
+			createdAt: meta.startTime,
+			modifiedAt: meta.modifiedTime,
+			workingDirectory: meta.workingDirectory?.toString(),
+		};
+
+		this._stateManager.restoreSession(summary, turns);
+		this._logService.info(`[AgentService] Restored session ${sessionStr} with ${turns.length} turns`);
 	}
 
 	async fetchContent(uri: URI): Promise<IFetchContentResult> {
-		return this._sideEffects.handleFetchContent(uri.toString());
+		// Handle session-db: URIs that reference file-edit content stored
+		// in a per-session SQLite database.
+		const dbFields = parseSessionDbUri(uri.toString());
+		if (dbFields) {
+			return this._fetchSessionDbContent(dbFields);
+		}
+
+		try {
+			const content = await this._fileService.readFile(uri);
+			return {
+				data: content.value.toString(),
+				encoding: ContentEncoding.Utf8,
+				contentType: 'text/plain',
+			};
+		} catch (_e) {
+			throw new ProtocolError(AhpErrorCodes.NotFound, `Content not found: ${uri.toString()}`);
+		}
 	}
 
 	async writeFile(params: IWriteFileParams): Promise<IWriteFileResult> {
-		return this._sideEffects.handleWriteFile(params);
+		const fileUri = typeof params.uri === 'string' ? URI.parse(params.uri) : URI.revive(params.uri);
+		let content: VSBuffer;
+		if (params.encoding === ContentEncoding.Base64) {
+			content = decodeBase64(params.data);
+		} else {
+			content = VSBuffer.fromString(params.data);
+		}
+		try {
+			if (params.createOnly) {
+				await this._fileService.createFile(fileUri, content, { overwrite: false });
+			} else {
+				await this._fileService.writeFile(fileUri, content);
+			}
+			return {};
+		} catch (e) {
+			const code = toFileSystemProviderErrorCode(e as Error);
+			if (code === FileSystemProviderErrorCode.FileExists) {
+				throw new ProtocolError(AhpErrorCodes.AlreadyExists, `File already exists: ${fileUri.toString()}`);
+			}
+			if (code === FileSystemProviderErrorCode.NoPermissions) {
+				throw new ProtocolError(AhpErrorCodes.PermissionDenied, `Permission denied: ${fileUri.toString()}`);
+			}
+			throw new ProtocolError(AhpErrorCodes.NotFound, `Failed to write file: ${fileUri.toString()}`);
+		}
 	}
 
 	async shutdown(): Promise<void> {
@@ -232,6 +341,120 @@ export class AgentService extends Disposable implements IAgentService {
 	}
 
 	// ---- helpers ------------------------------------------------------------
+
+	/**
+	 * Reconstructs completed `ITurn[]` from a sequence of agent session
+	 * messages. Each user-message starts a new turn; the assistant message
+	 * closes it.
+	 */
+	private _buildTurnsFromMessages(
+		messages: readonly (IAgentMessageEvent | IAgentToolStartEvent | IAgentToolCompleteEvent)[],
+	): ITurn[] {
+		const turns: ITurn[] = [];
+		let currentTurn: {
+			id: string;
+			userMessage: { text: string };
+			responseParts: IResponsePart[];
+			pendingTools: Map<string, IAgentToolStartEvent>;
+		} | undefined;
+
+		const finalizeTurn = (turn: NonNullable<typeof currentTurn>, state: TurnState): void => {
+			turns.push({
+				id: turn.id,
+				userMessage: turn.userMessage,
+				responseParts: turn.responseParts,
+				usage: undefined,
+				state,
+			});
+		};
+
+		const startTurn = (id: string, text: string): NonNullable<typeof currentTurn> => ({
+			id,
+			userMessage: { text },
+			responseParts: [],
+			pendingTools: new Map(),
+		});
+
+		for (const msg of messages) {
+			if (msg.type === 'message' && msg.role === 'user') {
+				if (currentTurn) {
+					finalizeTurn(currentTurn, TurnState.Cancelled);
+				}
+				currentTurn = startTurn(msg.messageId, msg.content);
+			} else if (msg.type === 'message' && msg.role === 'assistant') {
+				if (!currentTurn) {
+					currentTurn = startTurn(msg.messageId, '');
+				}
+
+				if (msg.content) {
+					currentTurn.responseParts.push({
+						kind: ResponsePartKind.Markdown,
+						id: generateUuid(),
+						content: msg.content,
+					});
+				}
+
+				if (!msg.toolRequests || msg.toolRequests.length === 0) {
+					finalizeTurn(currentTurn, TurnState.Complete);
+					currentTurn = undefined;
+				}
+			} else if (msg.type === 'tool_start') {
+				currentTurn?.pendingTools.set(msg.toolCallId, msg);
+			} else if (msg.type === 'tool_complete') {
+				if (currentTurn) {
+					const start = currentTurn.pendingTools.get(msg.toolCallId);
+					currentTurn.pendingTools.delete(msg.toolCallId);
+
+					const tc: IToolCallCompletedState = {
+						status: ToolCallStatus.Completed,
+						toolCallId: msg.toolCallId,
+						toolName: start?.toolName ?? 'unknown',
+						displayName: start?.displayName ?? 'Unknown Tool',
+						invocationMessage: start?.invocationMessage ?? '',
+						toolInput: start?.toolInput,
+						success: msg.result.success,
+						pastTenseMessage: msg.result.pastTenseMessage,
+						content: msg.result.content,
+						error: msg.result.error,
+						confirmed: ToolCallConfirmationReason.NotNeeded,
+						_meta: start ? {
+							toolKind: start.toolKind,
+							language: start.language,
+						} : undefined,
+					};
+					currentTurn.responseParts.push({
+						kind: ResponsePartKind.ToolCall,
+						toolCall: tc,
+					});
+				}
+			}
+		}
+
+		if (currentTurn) {
+			finalizeTurn(currentTurn, TurnState.Cancelled);
+		}
+
+		return turns;
+	}
+
+	private async _fetchSessionDbContent(fields: ISessionDbUriFields): Promise<IFetchContentResult> {
+		const sessionUri = URI.parse(fields.sessionUri);
+		const ref = this._sessionDataService.openDatabase(sessionUri);
+		try {
+			const content = await ref.object.readFileEditContent(fields.toolCallId, fields.filePath);
+			if (!content) {
+				throw new ProtocolError(AhpErrorCodes.NotFound, `File edit not found: toolCallId=${fields.toolCallId}, filePath=${fields.filePath}`);
+			}
+			const bytes = fields.part === 'before' ? content.beforeContent : content.afterContent;
+			return {
+				data: new TextDecoder().decode(bytes),
+				encoding: ContentEncoding.Utf8,
+				contentType: 'text/plain',
+			};
+		} finally {
+			ref.dispose();
+		}
+	}
 
 	private _findProviderForSession(session: URI | string): IAgent | undefined {
 		const key = typeof session === 'string' ? session : session.toString();
