+++ src/vs/platform/agentHost/test/node/agentService.test.ts
@@ -4,14 +4,18 @@
  *--------------------------------------------------------------------------------------------*/
 
 import assert from 'assert';
+import { VSBuffer } from '../../../../base/common/buffer.js';
 import { DisposableStore, toDisposable } from '../../../../base/common/lifecycle.js';
+import { Schemas } from '../../../../base/common/network.js';
 import { URI } from '../../../../base/common/uri.js';
 import { ensureNoDisposablesAreLeakedInTestSuite } from '../../../../base/test/common/utils.js';
 import { NullLogService } from '../../../log/common/log.js';
 import { FileService } from '../../../files/common/fileService.js';
+import { InMemoryFileSystemProvider } from '../../../files/common/inMemoryFilesystemProvider.js';
 import { AgentSession } from '../../common/agentService.js';
 import { ISessionDataService } from '../../common/sessionDataService.js';
 import { ActionType, IActionEnvelope } from '../../common/state/sessionActions.js';
+import { ResponsePartKind, SessionLifecycle, ToolCallConfirmationReason, ToolCallStatus, ToolResultContentType, TurnState, type IMarkdownResponsePart, type IToolCallCompletedState, type IToolCallResponsePart } from '../../common/state/sessionState.js';
 import { AgentService } from '../../node/agentService.js';
 import { MockAgent } from './mockAgent.js';
 
@@ -20,8 +24,9 @@ suite('AgentService (node dispatcher)', () => {
 	const disposables = new DisposableStore();
 	let service: AgentService;
 	let copilotAgent: MockAgent;
+	let fileService: FileService;
 
-	setup(() => {
+	setup(async () => {
 		const nullSessionDataService: ISessionDataService = {
 			_serviceBrand: undefined,
 			getSessionDataDir: () => URI.parse('inmemory:/session-data'),
@@ -30,7 +35,14 @@ suite('AgentService (node dispatcher)', () => {
 			deleteSessionData: async () => { },
 			cleanupOrphanedData: async () => { },
 		};
-		service = disposables.add(new AgentService(new NullLogService(), disposables.add(new FileService(new NullLogService())), nullSessionDataService));
+		fileService = disposables.add(new FileService(new NullLogService()));
+		disposables.add(fileService.registerProvider(Schemas.inMemory, disposables.add(new InMemoryFileSystemProvider())));
+
+		// Seed a directory for browseDirectory tests
+		await fileService.createFolder(URI.from({ scheme: Schemas.inMemory, path: '/testDir' }));
+		await fileService.writeFile(URI.from({ scheme: Schemas.inMemory, path: '/testDir/file.txt' }), VSBuffer.fromString('hello'));
+
+		service = disposables.add(new AgentService(new NullLogService(), fileService, nullSessionDataService));
 		copilotAgent = new MockAgent('copilot');
 		disposables.add(toDisposable(() => copilotAgent.dispose()));
 	});
@@ -224,6 +236,110 @@ suite('AgentService (node dispatcher)', () => {
 
 			await service.shutdown();
 			assert.ok(copilotShutdown);
+		});
+	});
+
+	// ---- restoreSession -------------------------------------------------
+
+	suite('restoreSession', () => {
+
+		test('restores a session with message history', async () => {
+			service.registerProvider(copilotAgent);
+			const session = await copilotAgent.createSession();
+			const sessions = await copilotAgent.listSessions();
+			const sessionResource = sessions[0].session;
+
+			copilotAgent.sessionMessages = [
+				{ type: 'message', session, role: 'user', messageId: 'msg-1', content: 'Hello', toolRequests: [] },
+				{ type: 'message', session, role: 'assistant', messageId: 'msg-2', content: 'Hi there!', toolRequests: [] },
+			];
+
+			await service.restoreSession(sessionResource);
+
+			const state = service.stateManager.getSessionState(sessionResource.toString());
+			assert.ok(state, 'session should be in state manager');
+			assert.strictEqual(state!.lifecycle, SessionLifecycle.Ready);
+			assert.strictEqual(state!.turns.length, 1);
+			assert.strictEqual(state!.turns[0].userMessage.text, 'Hello');
+			const mdPart = state!.turns[0].responseParts.find((p): p is IMarkdownResponsePart => p.kind === ResponsePartKind.Markdown);
+			assert.ok(mdPart);
+			assert.strictEqual(mdPart.content, 'Hi there!');
+			assert.strictEqual(state!.turns[0].state, TurnState.Complete);
+		});
+
+		test('restores a session with tool calls', async () => {
+			service.registerProvider(copilotAgent);
+			const session = await copilotAgent.createSession();
+			const sessions = await copilotAgent.listSessions();
+			const sessionResource = sessions[0].session;
+
+			copilotAgent.sessionMessages = [
+				{ type: 'message', session, role: 'user', messageId: 'msg-1', content: 'Run a command', toolRequests: [] },
+				{ type: 'message', session, role: 'assistant', messageId: 'msg-2', content: 'I will run a command.', toolRequests: [{ toolCallId: 'tc-1', name: 'shell' }] },
+				{ type: 'tool_start', session, toolCallId: 'tc-1', toolName: 'shell', displayName: 'Shell', invocationMessage: 'Running command...' },
+				{ type: 'tool_complete', session, toolCallId: 'tc-1', result: { success: true, pastTenseMessage: 'Ran command', content: [{ type: ToolResultContentType.Text, text: 'output' }] } },
+				{ type: 'message', session, role: 'assistant', messageId: 'msg-3', content: 'Done!', toolRequests: [] },
+			];
+
+			await service.restoreSession(sessionResource);
+
+			const state = service.stateManager.getSessionState(sessionResource.toString());
+			assert.ok(state);
+			const turn = state!.turns[0];
+			const toolCallParts = turn.responseParts.filter((p): p is IToolCallResponsePart => p.kind === ResponsePartKind.ToolCall);
+			assert.strictEqual(toolCallParts.length, 1);
+			const tc = toolCallParts[0].toolCall as IToolCallCompletedState;
+			assert.strictEqual(tc.status, ToolCallStatus.Completed);
+			assert.strictEqual(tc.toolCallId, 'tc-1');
+			assert.strictEqual(tc.confirmed, ToolCallConfirmationReason.NotNeeded);
+		});
+
+		test('flushes interrupted turns', async () => {
+			service.registerProvider(copilotAgent);
+			const session = await copilotAgent.createSession();
+			const sessions = await copilotAgent.listSessions();
+			const sessionResource = sessions[0].session;
+
+			copilotAgent.sessionMessages = [
+				{ type: 'message', session, role: 'user', messageId: 'msg-1', content: 'Interrupted', toolRequests: [] },
+				{ type: 'message', session, role: 'user', messageId: 'msg-2', content: 'Retried', toolRequests: [] },
+				{ type: 'message', session, role: 'assistant', messageId: 'msg-3', content: 'Answer', toolRequests: [] },
+			];
+
+			await service.restoreSession(sessionResource);
+
+			const state = service.stateManager.getSessionState(sessionResource.toString());
+			assert.ok(state);
+			assert.strictEqual(state!.turns.length, 2);
+			assert.strictEqual(state!.turns[0].state, TurnState.Cancelled);
+			assert.strictEqual(state!.turns[1].state, TurnState.Complete);
+		});
+
+		test('throws when session is not found on backend', async () => {
+			service.registerProvider(copilotAgent);
+			await assert.rejects(
+				() => service.restoreSession(AgentSession.uri('copilot', 'nonexistent')),
+				/Session not found on backend/,
+			);
+		});
+	});
+
+	// ---- browseDirectory ------------------------------------------------
+
+	suite('browseDirectory', () => {
+
+		test('throws when the directory does not exist', async () => {
+			await assert.rejects(
+				() => service.browseDirectory(URI.from({ scheme: Schemas.inMemory, path: '/nonexistent' })),
+				/Directory not found/,
+			);
+		});
+
+		test('throws when the target is not a directory', async () => {
+			await assert.rejects(
+				() => service.browseDirectory(URI.from({ scheme: Schemas.inMemory, path: '/testDir/file.txt' })),
+				/Not a directory/,
+			);
 		});
 	});
 });
